requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
blas=1.0=mkl
bottleneck=1.3.5=py39h67323c0_0
bzip2=1.0.8=h0d85af4_4
ca-certificates=2022.4.26=hecd8cb5_0
cairo=1.16.0=h0ab9d94_1001
certifi=2022.6.15=py39hecd8cb5_0
et_xmlfile=1.1.0=py39hecd8cb5_0
ffmpeg=4.3.2=h4dad6da_0
fontconfig=2.13.1=h1027ab8_1000
freetype=2.10.4=h4cff582_1
gettext=0.19.8.1=h7937167_1005
glib=2.68.4=he49afe7_0
glib-tools=2.68.4=he49afe7_0
gmp=6.2.1=h2e338ed_0
gnutls=3.6.13=h756fd2b_1
graphite2=1.3.14=he9d5cce_1
harfbuzz=2.8.1=ha8c618c_0
hdf5=1.10.6=nompi_haae91d6_101
icu=58.2=h0a44026_3
intel-openmp=2021.4.0=hecd8cb5_3538
jasper=1.900.1=h636a363_1006
jbig=2.1=h0d85af4_2003
jpeg=9e=hca72f7f_0
lame=3.100=h35c211d_1001
lerc=2.2.1=h046ec9c_0
libblas=3.9.0=12_osx64_mkl
libcblas=3.9.0=12_osx64_mkl
libcxx=12.0.0=h2f01273_0
libdeflate=1.7=h35c211d_5
libffi=3.3=hb1e8313_2
libgfortran=4.0.0=7_5_0_h1a10cd1_23
libgfortran4=7.5.0=h1a10cd1_23
libglib=2.68.4=hd556434_0
libiconv=1.16=haf1e3a3_0
liblapack=3.9.0=12_osx64_mkl
liblapacke=3.9.0=12_osx64_mkl
libopencv=4.5.2=py39h852ad08_1
libpng=1.6.37=ha441bb4_0
libprotobuf=3.16.0=hcf210ce_0
libtiff=4.3.0=h1167814_1
libwebp-base=1.2.2=h0d85af4_1
libxml2=2.9.9=hd80cff7_2
llvm-openmp=14.0.4=ha654fa7_0
lz4-c=1.9.3=he49afe7_1
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
ncurses=6.3=hca72f7f_3
nettle=3.6=hedd7734_0
numexpr=2.8.3=py39h2e5f0a9_0
numpy=1.22.3=py39h2e5f0a9_0
numpy-base=1.22.3=py39h3b1a694_0
opencv=4.5.2=py39h6e9494a_1
openh264=2.1.1=hfd3ada9_0
openpyxl=3.0.10=py39hca72f7f_0
openssl=1.1.1q=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39he9d5cce_0
pcre=8.45=he49afe7_0
pip=22.1.2=py39hecd8cb5_0
pixman=0.38.0=h01d97ff_1003
py-opencv=4.5.2=py39h71a6800_1
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.9.2=py39h23ab428_6
python=3.9.12=hdfd78df_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pytz=2022.1=py39hecd8cb5_0
qt=5.9.7=h468cd18_1
readline=8.1.2=hca72f7f_1
setuptools=61.2.0=py39hecd8cb5_0
sip=4.19.13=py39h23ab428_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.38.5=h707629a_0
tk=8.6.12=h5d9f67b_0
tzdata=2022a=hda174b7_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!161.3030=h0d85af4_1
xz=5.2.5=hca72f7f_1
zlib=1.2.12=h4dc903c_2
zstd=1.5.0=h582d3a0_0
